requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
aiosqlite==0.21.0
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20 